Update ECMAScript target version to es6
As a developer working in a Linux environment (specifically using Ubuntu 14.04), this change was needed to make the example work. Under the assumption that this change does not break the example for other environments (i.e. Windows), I recommend incorporating it into the project to allow developers in any environment to "quick start" their learning and prototyping in Angular 2.